unittest: fix the `_GetOutOf_testPartExecutor` exception kind. Fixes https://github.com/pytest-dev/pytest/issues/6947
`_GetOutOf_testPartExecutor` was introduced in
https://github.com/pytest-dev/pytest/commit/04f27d4eb469fb9c76fd2d100564a0f7d30028df
and allow to catch Exception from a unittest TestCase, but as it
inheritate from `KeyboardInterrupt` it is managed differently by
unittest (see.
https://github.com/python/cpython/blob/032de7324e30c6b44ef272cea3be205a3d768759/Lib/unittest/case.py#L61-L62)
as the `testPartExecutor` reraise it directly. This prevent unittest to
run the cleanup functions properly.

Since this change enable the teardown methods to be ran properly by
unittest, the explicit teardown is not required anymore.