Simplify positional arguments compatibility code in pytest.fixture()
The dynamic scope feature added in 10bf6aac76d5060a0db4a94871d6dcf0a1
necessitated some wrangling of arguments in pytest.fixture(). In
particular, it deprecated positional arguments in favor of keyword-only
arguments, while keeping backward compatibility.

The way it did this avoided some code duplication but ended up being
quite hard to follow and to annotate with types.

Replace it with some straightforward code, which is not very DRY but is
simple and easy to remove when the time comes.