runner: inline helpers into runtestprotocol
runtestprotocol() is an important function which calls all of
pytest_runtest_* hooks for a single test item.

Previously it was broken up into several helper functions, to avoid some
duplication and add some abstraction. But I think it makes the code a
bit harder to follow, due to indirection and having to jump around, adds
stackframes and makes things harder to grep for.

Let's inline them all so there is one straight function which can be
read in all its glory.

Interesting take in the context of game programming if interested:
http://number-none.com/blow/john_carmack_on_inlined_code.html