Add TestPath class and test_path fixture
Also change Testdir to be a thin wrapper which only delegates to TestPath. An
alternative would be to subclass and only overwrite the methods that return
Path to convert the return to py.path.local, but this was preferred so we don't
expose Path-returning objects by accident and to expose exactly what we want
in Testdir.