skipping: fix dynamic xfail mark added in runtest not respected
If a test runtest phase (not setup) dynamically adds a pytest.mark.xfail
mark to the item, it should be respected, but it wasn't. This regressed
in 3e6fe92b7ea3c120e8024a970bf37a7c6c137714 (not released).

Fix it by just always refreshing the mark if needed. This is mostly what
was done before but in a more roundabout way.