python_api: safe repr() for ApproxScalar
This also eliminates the explicit check for `complex` in __repr__(), as mypy
complains that a common subclass of `numbers.Complex` and `complex` would have
conflicting method signatures. It's not clear where the conflict comes from, but
in any case it's simple (and slightly more Pythonic) to skirt the issue by
treating all numbers with a nonzero imaginary component as complex, and
considering complex-typed, but real-valued, numbers to be real.