terminal: fix crash in header reporting when testpaths is used without cli arguments
Regressed in 6.1.0 in 62e249a1f934d1073c9a0167077e133c5e0f6270.
The `x` is an `str` but is expected to be a `pathlib.Path`. Not caught
by mypy because `config.getini()` returns `Any`.

The code was actually wrong before, even `py.path.local`'s `bestrelpath`
function expects a `py.path.local`, not an `str`. But it had some weird
`try ... except AttributeError` fallback which just returns the
argument, i.e. it was a no-op.