Clarify fixture order and provide images for visual learners
The documentation previously stated that fixture execution
order was dictated by the order fixtures were requested in, but
this isn't a very reliable method controlling fixture order and
can lead to some confusing situations. The only reliable way to
control fixture order is based on scope, fixture dependencies,
and autouse. This reflects that in the documentation, and
provides several examples and images to help clarify that.

It also includes numerous additions and a good amount of
restructuring to help inform new users of the terminology used
and what the types of fixtures are. Advanced examples were also
thrown in.

Co-authored-by: Bruno Oliveira <nicoddemus@gmail.com>
Co-authored-by: Ran Benita <ran@unusedvar.com>