Refactor internal scope handling by introducing Scope enum
This replaces the internal use of string literals and "scopenum"
to a proper Scope enum, which also centralizes ordering, getting
higher and lower scopes, etc. Another benefit
is that it helps the type checker by introducing a proper type,
and improves readability.

Added a trivial changelog as courtesy for plugin authors that used the
removed attribute.

This is _mostly_ an internal change, however due to historical
reasons, the API of the following _internal_ objects has
changed slightly:

* `CallSpec2`: `_arg2scopenum` renamed to `_arg2scope` for consistency.
* `FixtureRequest`:
  Previously contained a `scope: str` attribute.
  Changed attribute to `_scope: Scope`, with a read-only property `scope -> str`.
* `SubRequest.__init__` parameter changed from `_Scope` to `Scope`.